requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
python-jose
python-multipart
python-socketio
